% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVAIREVA.R
\name{ANOVAIREVA}
\alias{ANOVAIREVA}
\title{Launch 'ANOVAIREVA' Interface}
\usage{
ANOVAIREVA()
}
\value{
Nothing
}
\description{
ANOVAIREVA() loads interactive user interface built using R 'shiny'.
}
\details{
Start ANOVAIREVA

The interactive user interface is to provide an easy way for anova analysis and downloading relevant plots.
}
\examples{
if(interactive()){
library(rmarkdown)
ANOVAIREVA()
}
}
\keyword{ANOVAIREVA}
